requirements.txt
playwright>=1.45.0,<1.50.0
python-dotenv>=1.0.1,<2.0.0
DrissionPage
loguru>=0.7.2
requests>=2.31.0
ddddocr>=1.4.11
opencv-python-headless>=4.8.0.0
numpy>=1.24.0
lxml>=4.9.0
